requirements.txt
aistudio_sdk==0.3.5
annotated-types==0.7.0
anyio==4.10.0
bce-python-sdk==0.9.42
beautifulsoup4==4.13.4
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.3
exceptiongroup==1.3.0
filelock==3.19.1
fire==0.7.1
fonttools==4.59.1
fsspec==2025.7.0
future==1.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hf-xet==1.1.8
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.3
lxml==6.0.0
matplotlib==3.10.5
modelscope==1.29.0
mss==10.1.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.0
pandas==2.3.2
pillow==11.3.0
prettytable==3.16.0
protobuf==6.32.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.5
rfc3986==1.5.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
termcolor==3.1.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.13
